+++ sequenziale/src/sort_int.h
@@ -9,20 +9,20 @@
  * Pivot = data[alto] (last element)
  * Returns final pivot position
  */
-long partition_lomuto(long *data, long basso,long alto);
+int partition_lomuto(int *data, int basso,int alto);
 
 /**
  * Hoare partition scheme: bidirectional convergence
  * Pivot = data[basso] (first element, or pre-selected)
  * Returns pivot position
  */
-long partition_hoare(long *data,long basso,long alto);
+int partition_hoare(int *data,int basso,int alto);
 
-void quick_sort(long *data,long basso,long alto);
+void quick_sort(int *data,int basso,int alto);
 
-void select_median_of_3(long *data, long basso, long alto);
-void median_quick_sort(long *data,long basso,long alto);
+void select_median_of_3(int *data, int basso, int alto);
+void median_quick_sort(int *data,int basso,int alto);
 
-void merge(long* data,size_t low,size_t mid,size_t high);
+void merge(int* data,size_t low,size_t mid,size_t high);
 
-void merge_sort(long* data,size_t left,size_t right);
+void merge_sort(int* data,size_t left,size_t right);
